Pull request #39: Some fixes
Merge in ADGUARD-CORE-LIBS/vpn-libs-endpoint from fix/some_fixes to master

Squashed commit of the following:

commit 4abe912b5f00a943fc8171cf078af638276c6e47
Author: Sergei Gunchenko <s.gunchenko@adguard.com>
Date:   Thu Feb 16 14:21:43 2023 +0300

    fix

commit c1c59748a82fd9097fbbcf92a617e332f0c16546
Author: Sergei Gunchenko <s.gunchenko@adguard.com>
Date:   Thu Feb 16 13:15:47 2023 +0300

    fix

commit b4b1981f31b111f5bc19fc7360f1b7db41d9dd38
Author: Sergei Gunchenko <s.gunchenko@adguard.com>
Date:   Thu Feb 16 12:12:25 2023 +0300

    minor

commit 690cb5db3fdf925fd99893943871befd882fd559
Author: Sergei Gunchenko <s.gunchenko@adguard.com>
Date:   Thu Feb 16 12:00:12 2023 +0300

    fix bench docker image

commit 78950a866c971c0db1ae9aba8f8be4a5fab3c4b1
Author: Sergei Gunchenko <s.gunchenko@adguard.com>
Date:   Wed Feb 15 19:33:19 2023 +0300

    apply some linter suggestions

commit 68d8ccecbc35b864a0f2788cc4373e1710632618
Author: Sergei Gunchenko <s.gunchenko@adguard.com>
Date:   Wed Feb 15 19:03:59 2023 +0300

    upgrade deps

commit 91969f12a6f3995c26e02b55542c0cf436bdd053
Author: Sergei Gunchenko <s.gunchenko@adguard.com>
Date:   Wed Feb 15 18:35:33 2023 +0300

    upgrade docker image version

commit f59f08ffff6236b29705e8bdde2e3b5074cb982b
Author: Sergei Gunchenko <s.gunchenko@adguard.com>
Date:   Wed Feb 15 18:11:35 2023 +0300

    add tracing feature

commit 4010735f52167cdab49d41aa1d7f08deaca85212
Author: Sergei Gunchenko <s.gunchenko@adguard.com>
Date:   Wed Feb 15 17:37:35 2023 +0300

    fix indentation

commit 2c53ae1462a344e1da006d4447b8ffc5a7fbd311
Author: Sergei Gunchenko <s.gunchenko@adguard.com>
Date:   Wed Feb 15 16:43:30 2023 +0300

    Fix busy loop in http3 codec

commit 717cf6cfe7d107a33a7914cabfd25fbd6701dfc6
Author: Sergei Gunchenko <s.gunchenko@adguard.com>
Date:   Tue Feb 14 11:42:07 2023 +0300

    apply formatter

commit 29d42ee2f735c635735c8762ec779f1980b4fedb
Author: Sergei Gunchenko <s.gunchenko@adguard.com>
Date:   Thu Feb 9 18:12:13 2023 +0300

    fix quic parameters names

commit 3c5652233293e53931de8dad40f95129ece3ddf6
Author: Sergei Gunchenko <s.gunchenko@adguard.com>
Date:   Thu Feb 9 14:14:20 2023 +0300

    apply linter suggestions

commit 73112b9dd8336f2baac6501b531ec50aadedf322
Author: Sergei Gunchenko <s.gunchenko@adguard.com>
Date:   Wed Feb 8 16:27:24 2023 +0300

    upgrade deps

commit dab47fb8e0c8f9cb44cff7a194b724b1b76c504e
Author: Sergei Gunchenko <s.gunchenko@adguard.com>
Date:   Wed Feb 8 13:37:12 2023 +0300

    bench: fix single host script

+++ lib/src/authentication/radius/ms_chap_v2.rs
@@ -362,7 +362,7 @@ fn des_parity_key(key_in: &[u8; 7]) -> [u8; 8] {
     let mut out = [0; 8];
 
     for (i, x) in key_in.iter().cloned().enumerate() {
-        out[i] = ((x >> i) | c_next | 1) as u8;
+        out[i] = (x >> i) | c_next | 1;
         c_next = x << (7 - i);
     }
